Refactor: Member - property 관리
