+++ src/main/java/com/todorian/member/command/application/service/MemberStatusService.java
@@ -3,9 +3,8 @@ package com.todorian.member.command.application.service;
 import com.todorian._core.error.exception.Exception400;
 import com.todorian.member.command.application.dto.MemberRequestDTO;
 import com.todorian.member.command.domain.model.Member;
-import com.todorian.member.command.domain.model.Status;
+import com.todorian.member.command.domain.model.property.Status;
 import com.todorian.member.command.domain.repository.MemberRepository;
-import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.crypto.password.PasswordEncoder;
