+++ src/main/java/com/todorian/member/command/domain/model/property/Authority.java
@@ -1,0 +1,5 @@
+package com.todorian.member.command.domain.model.property;
+
+public enum Authority {
+    ADMIN, USER, GUEST, NONE
+}
