+++ src/main/java/com/todorian/member/command/infrastructure/service/MemberOAuthServiceImpl.java
@@ -3,10 +3,10 @@ package com.todorian.member.command.infrastructure.service;
 import com.todorian._core.error.exception.Exception500;
 import com.todorian._core.jwt.JWTTokenProvider;
 import com.todorian.member.command.application.dto.MemberResponseDTO;
-import com.todorian.member.command.domain.model.Authority;
+import com.todorian.member.command.domain.model.property.Authority;
 import com.todorian.member.command.domain.model.Member;
-import com.todorian.member.command.domain.model.SocialType;
-import com.todorian.member.command.domain.model.Status;
+import com.todorian.member.command.domain.model.property.SocialType;
+import com.todorian.member.command.domain.model.property.Status;
 import com.todorian.member.command.domain.model.property.KakaoProperties;
 import com.todorian.member.command.domain.repository.MemberRepository;
 import com.todorian.member.command.domain.service.MemberOAuthService;
